created separate registerTRAR unit
The registration code was in RAR.pas.
This causes a problem when compiling a 32-bit application which "uses" RAR.pas and RAR_DLL.pas, as the Delphi IDE refuses to find DesignEditors and DesignIntf.
This is corrected by having an entirely separate registration unit which is only included in the package project.